Encapsulate Presenter logic by Store
