+++ lib/masamune/transform.rb
@@ -22,26 +22,6 @@
 
 module Masamune
   module Transform
-    require 'masamune/transform/operator'
-
-    require 'masamune/transform/define_table'
-    require 'masamune/transform/define_schema'
-    require 'masamune/transform/denormalize_table'
-
-    require 'masamune/transform/bulk_upsert'
-    require 'masamune/transform/insert_reference_values'
-
-    require 'masamune/transform/stage_dimension'
-    require 'masamune/transform/stage_fact'
-
-    require 'masamune/transform/load_dimension'
-    require 'masamune/transform/load_fact'
-
-    require 'masamune/transform/snapshot_dimension'
-    require 'masamune/transform/deduplicate_dimension'
-    require 'masamune/transform/consolidate_dimension'
-    require 'masamune/transform/relabel_dimension'
-
-    require 'masamune/transform/rollup_fact'
+    Dir["#{File.dirname(__FILE__)}/transform/*.rb"].each { |f| require f }
   end
 end
