+++ lib/masamune/transform/hive/define_schema.hql.erb
