+++ lib/masamune/transform/hive/define_table.hql.erb
