+++ lib/masamune/transform/hive/denormalize_table.hql.erb
