+++ lib/masamune/transform/postgres/bulk_upsert.psql.erb
