+++ lib/masamune/transform/postgres/define_foreign_key.psql.erb
