+++ lib/masamune/transform/postgres/define_index.psql.erb
