+++ lib/masamune/transform/postgres/define_inheritance.psql.erb
