+++ lib/masamune/transform/postgres/define_schema.psql.erb
