+++ lib/masamune/transform/postgres/define_table.psql.erb
