+++ lib/masamune/transform/postgres/define_unique.psql.erb
