+++ lib/masamune/transform/postgres/denormalize_table.psql.erb
