+++ lib/masamune/transform/postgres/denormalize_table.rb
@@ -20,15 +20,6 @@
 #  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 #  THE SOFTWARE.
 
-module Masamune::Transform
-  module DefineTable
-    extend ActiveSupport::Concern
-
-    def define_table(target, files = [], section = nil)
-      return if target.implicit
-      Operator.new(__method__, target: target, files: files, section: section).tap do |operator|
-        logger.debug("#{target.id}\n" + operator.to_s) if target.debug
-      end
-    end
-  end
+module Masamune::Transform::Postgres
+  DenormalizeTable = Masamune::Transform::Common::DenormalizeTable
 end
