+++ lib/masamune/transform/postgres/insert_reference_values.rb
@@ -1,0 +1,67 @@
+#  The MIT License (MIT)
+#
+#  Copyright (c) 2014-2015, VMware, Inc. All Rights Reserved.
+#
+#  Permission is hereby granted, free of charge, to any person obtaining a copy
+#  of this software and associated documentation files (the "Software"), to deal
+#  in the Software without restriction, including without limitation the rights
+#  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+#  copies of the Software, and to permit persons to whom the Software is
+#  furnished to do so, subject to the following conditions:
+#
+#  The above copyright notice and this permission notice shall be included in
+#  all copies or substantial portions of the Software.
+#
+#  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+#  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+#  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+#  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+#  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+#  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+#  THE SOFTWARE.
+
+module Masamune::Transform::Postgres
+  class InsertReferenceValues
+    include Masamune::Transform::BulkUpsert
+
+    def initialize(options = {})
+      @target   = options[:target]
+      @source   = options[:source]
+    end
+
+    def locals
+      { target: target, source: @source, helper: self }
+    end
+
+    def target
+      TargetPresenter.new(@target)
+    end
+
+    private
+
+    class TargetPresenter < SimpleDelegator
+      include Masamune::LastElement
+
+      def insert_columns(source)
+        source.shared_columns(stage_table).map { |_, columns| columns.first.name }
+      end
+
+      def insert_values(source)
+        source.shared_columns(stage_table).map do |column, _|
+          if column.adjacent.try(:default)
+            "COALESCE(#{column.name}, #{column.adjacent.sql_value(column.adjacent.default)})"
+          else
+            column.name
+          end
+        end
+      end
+      method_with_last_element :insert_values
+
+      def insert_constraints(source)
+        source.shared_columns(stage_table).reject { |column, _| column.null || column.default || column.adjacent.try(:default) }.map { |column, _| "#{column.name} IS NOT NULL"}
+      end
+      method_with_last_element :insert_constraints
+    end
+  end
+end
+
