+++ lib/masamune/transform/postgres/relabel_dimension.psql.erb
