+++ lib/masamune/transform/postgres/replace_table.psql.erb
