+++ lib/masamune/transform/postgres/rollup_fact.psql.erb
