+++ lib/masamune/transform/postgres/snapshot_dimension.psql.erb
