+++ lib/masamune/transform/postgres/stage_dimension.psql.erb
