+++ lib/masamune/transform/postgres/stage_fact.psql.erb
