+++ lib/masamune/transform/rollup_fact.rb
@@ -27,92 +27,7 @@ module Masamune::Transform
     def rollup_fact(source, target, date)
       raise ArgumentError, "#{source.name} must have date_column to rollup" unless source.date_column
       raise ArgumentError, "#{target.name} must have date_column to rollup" unless target.date_column
-      Operator.new __method__, source: source.partition_table(date), target: target.partition_table(date), presenters: { postgres: Postgres }
-    end
-
-    private
-
-    class Postgres < SimpleDelegator
-      include Masamune::LastElement
-
-      def insert_columns(source)
-        values = []
-        shared_columns(source).values.map do |columns|
-          column = columns.first
-          next if column.id == :last_modified_at
-          next if column.auto_reference
-          values << column.name
-        end
-        measures.each do |_ ,measure|
-          values << measure.name
-        end
-        values << time_key.name
-        values.compact
-      end
-
-      def insert_values(source)
-        values = []
-        values << calculated_date_key(source)
-        shared_columns(source).values.map do |columns|
-          column = columns.first
-          next unless column.reference
-          next if column.reference.type == :date
-          next if column.auto_reference
-          values << column.qualified_name
-        end
-        source.measures.each do |_ ,measure|
-          values << measure.aggregate_value
-        end
-        values << calculated_time_key(source)
-        values
-      end
-      method_with_last_element :insert_values
-
-      def join_conditions(source)
-        {
-          source.date_column.reference.name => [
-            "#{source.date_column.reference.surrogate_key.qualified_name} = #{source.date_column.qualified_name}"
-          ]
-        }
-      end
-
-      def group_by(source)
-        group_by = []
-        group_by << calculated_date_key(source)
-        shared_columns(source).values.map do |columns|
-          column = columns.first
-          next unless column.reference
-          next if column.reference.type == :date
-          next if column.auto_reference
-          group_by << column.qualified_name
-        end
-        group_by << calculated_time_key(source)
-        group_by
-      end
-      method_with_last_element :group_by
-
-      private
-
-      def calculated_date_key(source)
-        case grain
-        when :hourly, :daily
-          "#{source.date_column.qualified_name}"
-        when :monthly
-          "to_char(date_trunc('month',#{source.date_column.qualified_name}::text::date),'YYYYMMDD')::integer"
-        end
-      end
-
-      def calculated_time_key(source)
-        case grain
-        when :hourly
-          "(#{source.time_key.qualified_name} - (#{source.time_key.qualified_name} % #{1.hour.seconds}))"
-        when :daily
-          "extract(EPOCH from #{source.date_column.qualified_name}::text::date)"
-        when :monthly
-          "extract(EPOCH from date_trunc('month',#{source.date_column.qualified_name}::text::date))"
-        end
-      end
-
+      Operator.new __method__, source: source.partition_table(date), target: target.partition_table(date)
     end
   end
 end
