+++ lib/masamune/transform/stage_fact.rb
@@ -25,102 +25,7 @@ module Masamune::Transform
     extend ActiveSupport::Concern
 
     def stage_fact(source, target, date)
-      Operator.new(__method__, source: source, target: target, date: date, presenters: { postgres: Postgres })
-    end
-
-    private
-
-    class Postgres < SimpleDelegator
-      include Masamune::LastElement
-
-      def insert_columns(source)
-        shared_columns(source).values.map do |columns|
-          column = columns.first
-          if reference = column.reference
-            reference.foreign_key_name
-          else
-            column.name
-          end
-        end
-      end
-
-      def insert_values(source)
-        shared_columns(source).values.map do |columns|
-          column = columns.first
-          if !column.degenerate? && reference = column.reference
-            reference.surrogate_key.qualified_name(column.reference.label)
-          else
-            column.qualified_name
-          end
-        end
-      end
-      method_with_last_element :insert_values
-
-      def join_alias(reference)
-        reference.label ? "#{reference.name} AS #{[reference.label, reference.name].compact.join('_')}" : reference.name
-      end
-
-      def join_conditions(source)
-        join_columns = shared_columns(source).values.flatten
-        join_columns = join_columns.select { |column| column.reference }
-        join_columns = join_columns.group_by { |column| column.reference }
-
-        dependencies = Masamune::TopologicalHash.new
-        conditions = Hash.new { |h,k| h[k] = [] }
-        join_columns.each do |reference, columns|
-          reference_name = join_alias(reference)
-          columns.each do |column|
-            next if column.degenerate?
-            dependencies[reference_name] ||= []
-            cross_references = cross_references(column)
-
-            coalesce_values = []
-
-            if cross_references.any?
-              dependencies[reference_name] += cross_references.map { |reference, _| join_alias(reference) }
-              coalesce_values << cross_references.map { |reference, column| column.qualified_name(reference.label) }
-            end
-
-            column.reference.auto_surrogate_keys.each do |auto_surrogate_key|
-              next unless auto_surrogate_key.default
-              conditions[reference_name] << "#{auto_surrogate_key.qualified_name(reference.label)} = #{auto_surrogate_key.default}"
-            end if column.reference
-
-            if column.reference && !column.reference.default.nil? && column.adjacent.natural_key
-              coalesce_values << column.reference.default(column.adjacent)
-            elsif column.adjacent && !column.adjacent.default.nil?
-              coalesce_values << column.adjacent.sql_value(column.adjacent.default)
-            end
-
-            conditions[reference_name] << (coalesce_values.any? ?
-              "#{column.foreign_key_name} = COALESCE(#{column.qualified_name}, #{coalesce_values.join(', ')})" :
-              "#{column.foreign_key_name} = #{column.qualified_name}")
-          end
-
-          if reference.type == :two || reference.type == :four
-            join_key_a = "TO_TIMESTAMP(#{source.time_key.qualified_name}) BETWEEN #{reference.start_key.qualified_name(reference.label)} AND COALESCE(#{reference.end_key.qualified_name(reference.label)}, 'INFINITY')"
-            join_key_b = "TO_TIMESTAMP(#{source.time_key.qualified_name}) < #{reference.start_key.qualified_name(reference.label)} AND #{reference.version_key.qualified_name(reference.label)} = 1"
-            conditions[reference_name] << "((#{join_key_a}) OR (#{join_key_b}))"
-          end
-
-          conditions[reference_name].uniq!
-        end
-        conditions.slice(*dependencies.tsort)
-      end
-
-      private
-
-      def cross_references(column)
-        return {} unless column.natural_key || column.adjacent.try(:natural_key)
-        {}.tap do |result|
-          column.reference.through.each do |reference_id|
-            reference = references[reference_id]
-            if reference.columns[column.id]
-              result[reference] = reference.columns[column.id]
-            end
-          end
-        end
-      end
+      Operator.new(__method__, source: source, target: target, date: date)
     end
   end
 end
